fixes to router organization
