+++ app/api.py
@@ -1,20 +1,19 @@
 import os
 
+from app.routers import (mentor_router,
+                         mentee_router,
+                         feedback_router,
+                         graph_router,
+                         model_router)
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
 from app.data import MongoDB
 from app.model import MatcherSortSearch
-from app.routers import (mentor_router,
-                         mentee_router,
-                         data_router,
-                         model_router,
-                         graph_router)
-
 
 API = FastAPI(
     title='Underdog Devs DS API',
-    version="0.49.1",
+    version="0.49.2",
     docs_url='/',
 )
 
@@ -27,10 +26,6 @@ API.add_middleware(
     allow_methods=["*"],
     allow_headers=["*"],
 )
-
-@API.post("/info", tags=["General Operations"])
-async def info():
-    return {"info": "API Information"}
 
 
 @API.get("/version")
@@ -46,5 +41,5 @@ async def collections():
     return {"result": API.db.get_database_info()}
 
 
-for router in (mentor_router, mentee_router, data_router, model_router, graph_router):
+for router in (mentor_router, mentee_router, feedback_router, graph_router, model_router):
     API.include_router(router.Router)
